Dashboard, authContext y dashboardContext
Dashboard component functionality moved to dashboardContext.
authContext has some lines commented because I'm in the way to implement
strong types contexts and returns.
dashboardContext has now the main functionality handling states
that the Dashboard components needs such as the user document and
the registers collection to render the main UI.