Contexts
Both contexts files now have correct types declarations.
Added four new functions to DashboardContext:

1_ buildRegisterSchema: receives the parameters to build an schema to
be used as a register and returns it.
2_ addRegisterToLocalUserData: receives a register schema as an input
and modifies userData state adding the new register locally.
3_ addRegisterToFirestore: receives a register schema as an input and
adds it to the firestore register collection of the signed in user.
4_ addRegister: receives Formik Values (later this type is going to be
changed to newRegisterValuesInterface - an already exported interface)
and executes buildRegisterSchema, addRegisterToLocalUserData and
addRegistertoFirebase in that order. Later this function is going to be
modified so it can catch errors and show them in the add register modal.