Modified Dashboard , created two new components
Dashboard component is now cleared of old comments and now returns one
new component: Registers.

Created two new components: Registers and RegisterField.

Registers: container of all RegisterFields, manages the listing of
RegisterFields depending of the userData.
RegisterField: standart component for listing one register.

Both these new components are very basic as they were made simply for
cheking the correct working of the DashboardContext functionality and
the new structure of it's values (userData).
They will be improved later on as they are also mean to work as the
foundation for the former UI components that make the Dashboard.