addRegister changes
Modified the order of the functions, now the new registers gets sent to
the firestore to retrieve it's ID in the collection and combining the
new register with it's firestore id so we can use this id (listed as a
"key" in the JSON) as a key in the array map. This will be useful when
building the deleteRegister function.