Added cloud functions and react-select
The project will now use Firebase cloud functions to manage a crucial
aspect of the application: the exchange rates.

A new package has been added: react-select and it's corresponding types.
This will help create the UI for the dropdown selection of the symbols
and facilitate the development of the remaining modals.